Refactor - Refactor Guardian
Guardian - Rename to GuardianError
CHANGE - Simplify guardian error generation
ADD - Gitignore
