+++ endpoint/HTTPErrors.ts
@@ -1,9 +1,20 @@
-import { HttpError, Status } from "../dependencies.ts";
+import { createHttpError } from "https://deno.land/std@0.150.0/http/http_errors.ts";
+import { Status } from "../dependencies.ts";
 
-export class ResourceNotFound extends HttpError {
-  constructor(message: string) {
-    super(message, {
-      expose: true,
-    });
-  }
-}
+export const resourceNotFound = (message: string) => {
+  return createHttpError(Status.NotFound, message);
+};
+
+// export class ResourceNotFound extends HttpError {
+
+//   constructor(message: string) {
+//     super(message, {
+//       expose: true,
+//     });
+//   }
+// }
+
+// const a = createHttpError(Status.NotFound, "Not found", {expose: true});
+// console.log(a.status)
+// console.log(a.message)
+// console.log(a.headers)
