+++ guardian/Guardians/DateGuardian.ts
@@ -3,12 +3,12 @@ import { BaseGuardian } from "../BaseGuardian.ts";
 import { StringGuardian } from "./StringGuardian.ts";
 import { NumberGuardian } from "./NumberGuardian.ts";
 
-import { type } from "../utils.ts";
+import { type } from "../utils/mod.ts";
 import type {
   FunctionParameters,
   FunctionType,
   GuardianProxy,
-} from "../types.ts";
+} from "../types/mod.ts";
 
 /**
  * DateGuardian
