+++ guardian/Guardians/ObjectGuardian.ts
@@ -1,6 +1,6 @@
 import { BaseGuardian } from "../BaseGuardian.ts";
-import { type } from "../utils.ts";
-import type { FunctionParameters, FunctionType } from "../types.ts";
+import { type } from "../utils/mod.ts";
+import type { FunctionParameters, FunctionType } from "../types/mod.ts";
 
 /**
  * ObjectGuardian
