+++ guardian/Guardians/StringGuardian.ts
@@ -1,10 +1,10 @@
 import { BaseGuardian } from "../BaseGuardian.ts";
-import { type } from "../utils.ts";
+import { type } from "../utils/mod.ts";
 import type {
   FunctionParameters,
   FunctionType,
   GuardianProxy,
-} from "../types.ts";
+} from "../types/mod.ts";
 
 /**
  * StringGuardian
