+++ guardian/error/mod.ts
@@ -1,0 +1,2 @@
+export { makeError } from "./makeError.ts";
+export { GuardianError } from "./GuardianError.ts";
