+++ guardian/types/Object.ts
@@ -1,0 +1,3 @@
+export type ObjectProperty = string | number | symbol;
+
+export type ObjectPath = Array<ObjectProperty>;
