+++ guardian/types/Primitive.ts
@@ -1,0 +1,9 @@
+export type Primitive =
+  | string
+  | number
+  | symbol
+  | boolean
+  | null
+  | undefined
+  | void
+  | bigint;
