+++ guardian/types/Typeof.ts
@@ -1,0 +1,11 @@
+export type Typeof = {
+  string: string;
+  number: number;
+  // deno-lint-ignore ban-types
+  object: object;
+  boolean: boolean;
+  symbol: symbol;
+  bigint: bigint;
+  undefined: undefined;
+  date: Date;
+};
