+++ guardian/utils/isPromiseLike.ts
@@ -1,0 +1,3 @@
+export function isPromiseLike(value: unknown): value is PromiseLike<unknown> {
+  return !!value && typeof (value as PromiseLike<unknown>).then === "function";
+}
